# api-gateway/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: RetailHub API
  description: |
    Event-Driven платформа диспетчеризации консультантов в ритейле.
    Клиент сканирует QR-код → создаётся заявка → консультант берёт заявку.

    ## Аутентификация
    Закрытые эндпоинты требуют JWT Bearer-токен:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Роли
    - **MANAGER** — управление магазином, отделами, консультантами, аналитика
    - **CONSULTANT** — приём и обработка заявок, управление сменами

    ## Публичные эндпоинты (без токена)
    - Сканирование QR-кода
    - Создание заявки
    - Polling статуса заявки клиентом
    - Напоминание консультанту
    - Смена консультанта
  version: 1.0.0
  contact:
    name: RetailHub Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: http://83.147.255.205:8180/api/v1
    description: Production (Direct IP)
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Role:
      type: string
      enum: [MANAGER, CONSULTANT]
    UserStatus:
      type: string
      enum: [OFFLINE, ACTIVE, BUSY]
      description: |
        OFFLINE — не на смене, ACTIVE — свободен, BUSY — обслуживает клиента
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        store_id:
          type: string
          format: uuid
        phone_number:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        current_status:
          $ref: "#/components/schemas/UserStatus"
        departments:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentShort"
          description: Отделы, в которых сотрудник компетентен
        created_at:
          type: string
          format: date-time
    DepartmentShort:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    CreateUserRequest:
      type: object
      required: [phone_number, password, first_name, last_name, role]
      properties:
        phone_number:
          type: string
          description: Номер телефона для логина. Уникальный.
          example: "+79991234567"
        password:
          type: string
          format: password
          minLength: 6
        first_name:
          type: string
          maxLength: 100
          example: Иван
        last_name:
          type: string
          maxLength: 100
          example: Петров
        role:
          $ref: "#/components/schemas/Role"
        department_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Отделы для назначения компетенций
    UpdateUserRequest:
      type: object
      properties:
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
    AssignDepartmentsRequest:
      type: object
      required: [department_ids]
      properties:
        department_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Список отделов (полная замена текущих)

    # --- Store ---
    Store:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        timezone:
          type: string
          example: Europe/Moscow
        created_at:
          type: string
          format: date-time
    UpdateStoreRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        timezone:
          type: string
          example: Europe/Moscow
    CreateStoreRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
          example: "Main Store"
        address:
          type: string
          example: "ул. Ленина, 1"
        timezone:
          type: string
          example: Europe/Moscow

    # --- Department ---
    Department:
      type: object
      properties:
        id:
          type: string
          format: uuid
        store_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          description: Краткое описание отдела, показывается клиенту на экране ожидания
          example: Ноутбуки, планшеты, аксессуары
        created_at:
          type: string
          format: date-time
    CreateDepartmentRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
          example: Телевизоры
        description:
          type: string
          description: Краткое описание отдела (опционально)
          example: Смарт-телевизоры, проекторы, аксессуары
    UpdateDepartmentRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
          description: Краткое описание отдела (опционально)

    # --- QR Code ---
...
    Shift:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        user_name:
          type: string
          description: "Имя Фамилия"
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          description: NULL = смена ещё активна

    # --- Notification ---
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: REQUEST_ALREADY_ASSIGNED
        message:
          type: string
          example: Заявка уже взята другим консультантом
        timestamp:
          type: string
          format: date-time
    PageResponse:
      type: object
      properties:
        content:
          type: array
          items: {}
        page:
          type: integer
        size:
          type: integer
        total_elements:
          type: integer
        total_pages:
          type: integer

# PATHS
paths:
...
  /users:
    post:
      tags: [Users]
      summary: Создать аккаунт сотрудника
      description: Менеджер создаёт аккаунт в своём магазине. Можно сразу назначить отделы.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Валидация (телефон занят)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Нет прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags: [Users]
      summary: Список сотрудников магазина
      parameters:
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/Role"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Список
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageResponse"
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: "#/components/schemas/UserProfile"
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Users]
      summary: Профиль сотрудника
      responses:
        "200":
          description: Профиль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags: [Users]
      summary: Обновить сотрудника
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

    delete:
      tags: [Users]
      summary: Удалить сотрудника
      description: Полное удаление из БД. Связанные смены удаляются каскадно.
      responses:
        "204":
          description: Удалён
  /users/{userId}/departments:
    put:
      tags: [Users]
      summary: Назначить отделы (компетенции)
      description: Полная замена списка отделов.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignDepartmentsRequest"
      responses:
        "200":
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # --- STORES (MANAGER) ---
  /stores:
    post:
      tags: [Stores]
      summary: Создать магазин
      description: Только MANAGER. Создаёт новый магазин. timezone по умолчанию Europe/Moscow.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreRequest"
      responses:
        "201":
          description: Магазин создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "403":
          description: Нет прав (не MANAGER)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stores/my:
    get:
      tags: [Stores]
      summary: Мой магазин
      responses:
        "200":
          description: Магазин
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"

    put:
      tags: [Stores]
      summary: Обновить магазин
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoreRequest"
      responses:
        "200":
          description: Обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /stores/my/departments:
    get:
      tags: [Stores]
      summary: Список отделов
      responses:
        "200":
          description: Отделы
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"

    post:
      tags: [Stores]
      summary: Создать отдел
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentRequest"
      responses:
        "201":
          description: Создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{departmentId}:
    parameters:
      - name: departmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Stores]
      summary: Получить отдел по ID
      responses:
        "200":
          description: Отдел
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "404":
          description: Отдел не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags: [Stores]
      summary: Обновить отдел
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDepartmentRequest"
      responses:
        "200":
          description: Обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"

    delete:
      tags: [Stores]
      summary: Удалить отдел
      description: Каскадно удалит QR-коды и привязки консультантов.
      responses:
        "204":
          description: Удалён

  # --- QR CODES ---
...
  /shifts/start:
    post:
      tags: [Shifts]
      summary: Начать смену (CONSULTANT)
      description: Создаёт запись смены. Консультант → ACTIVE.
      responses:
        "201":
          description: Смена начата
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
        "409":
          description: Уже на смене
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shifts/end:
    post:
      tags: [Shifts]
      summary: Закончить смену (CONSULTANT)
      description: Ставит ended_at. Консультант → OFFLINE.
      responses:
        "200":
          description: Смена завершена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
        "400":
          description: Нет активной смены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /shifts/my:
    get:
      tags: [Shifts]
      summary: Мои смены (CONSULTANT)
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Мои смены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shift"

  # --- NOTIFICATIONS ---
...